+++ src/meta/table.spec.ts
@@ -272,6 +272,7 @@ function mockEntry(tagName: string, stub = {}): MetaElement {
 			implicitClosed: [],
 			attributes: {},
 			deprecatedAttributes: [],
+			requiredAttributes: [],
 			permittedContent: [],
 			permittedDescendants: [],
 			permittedOrder: [],