+++ src/meta/validator.spec.ts
@@ -483,6 +483,7 @@ function mockEntry(tagName: string, stub = {}): MetaElement {
 			implicitClosed: [],
 			attributes: {},
 			deprecatedAttributes: [],
+			requiredAttributes: [],
 			permittedContent: [],
 			permittedDescendants: [],
 			permittedOrder: [],